pyproject.toml
[project]
name = "gui-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chroma>=0.2.0",
    "langchain[google-genai]>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-core>=0.3.75",
    "pypdf>=6.0.0",
    "streamlit>=1.49.0",
    "pysqlite3-binary",
]
